Switched buildout for virtualenv(wrapper)
Instead of buildout, I've switched this project to use streight virtualenv.
All the dependencies for running this project are specified in setup.py. The
dependencies for developing the porject, can be installed using pip:
'pip install -r requirements.txt'.

The virtualenv is installed with distribute instead of setuptools.

TODO: Need to write some instructions in the README and INSTALL files.