Using bundler to manage gemfile and dependencies.
Following railscasts about bundler and gem, I converted
the project to use bundler to manage the dependencies
(although not in runtime). All dependencies are specified
in the gemspec file, and bundler reads the list from
there (both development and runtime).

To avoid problems running bundler for the first time
I split the version module from the main RunSSHLib
module so there won't be loading error while reading
the specfile.

Also, It seems to work on 1.9.2 so I changed the
required ruby version.